format/mainPage_treeInfos 전역 모달 관리 재구성 /#11
