+++ src/components/Layout/index.jsx
@@ -1,18 +1,10 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import { Outlet } from 'react-router-dom'
-import { useSelector, useDispatch } from 'react-redux'
 import { NavBar } from '../common/NavBar'
-import { Container, Overlay } from './style'
-import { showAddressModal, showShareModal } from '../../store/slices/modalSlice'
+import { Container } from './style'
 // import useKakaoSharing from '../../hooks/useKakaoSharing'
 
 export const Layout = () => {
-  const dispatch = useDispatch()
-  const adModal = useSelector((state) => state.modal.addressShow)
-  const shareModal = useSelector((state) => state.modal.shareShow)
-
-  useEffect(() => {}, [adModal, shareModal])
-
   // 카카오 공유하기
   // useEffect(() => {
   //   const script = document.createElement('script')
@@ -32,8 +24,6 @@ export const Layout = () => {
     <Container>
       <Outlet />
       <NavBar />
-      {adModal && <Overlay onClick={() => dispatch(showAddressModal())} />}
-      {shareModal && <Overlay onClick={() => dispatch(showShareModal())} />}
     </Container>
   )
 }
