+++ src/components/Layout/style.js
@@ -8,13 +8,3 @@ export const Container = styled.div`
   transform: translateX(-50%);
   background-color: #fff;
 `
-export const Overlay = styled.div`
-  box-sizing: border-box;
-  position: fixed;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  /* background-color: rgba(0, 0, 0, 0.5); */
-  z-index: 10;
-`
